Move Menu to helix-view
