+++ helix-view/src/ui/mod.rs
@@ -1,7 +1,10 @@
 mod info;
+pub mod menu;
+pub mod prompt;
 mod spinner;
 mod text;
 
 pub use info::Info;
 pub use spinner::{ProgressSpinners, Spinner};
 pub use text::Text;
+pub use {menu::Item as MenuItem, menu::Menu};
